pyproject.toml
[project]
name = "biofoundry-cli"
version = "1.22.0"
description = "Biofoundry_CLI is your next CLI agent."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agent-client-protocol==0.8.0",
    "aiofiles>=24.0,<26.0",
    "aiohttp==3.13.3",
    "typer==0.21.1",
    "biofoundry_llm[contrib]==0.1.0",
    # loguru stays >=0.6.0 because notify-py (via batrachian-toad) caps it at <=0.6.0 on 3.14+.
    "loguru>=0.6.0,<0.8",
    "prompt-toolkit==3.0.52",
    "pillow==12.1.0",
    "pyyaml==6.0.3",
    "rich==14.2.0",
    "ripgrepy==2.2.0",
    "streamingjson==0.0.5",
    "trafilatura==2.0.0",
    # lxml is used by trafilatura/htmldate/justext; keep pinned for binary wheels.
    "lxml==6.0.2",
    "tenacity==9.1.2",
    "fastmcp==2.12.5",
    "pydantic==2.12.5",
    "httpx[socks]==0.28.1",
    "biofoundry_os==0.1.0",
    "batrachian-toad==0.5.23; python_version >= \"3.14\"",
    "tomlkit==0.14.0",
    "jinja2==3.1.6",
    "pyobjc-framework-cocoa>=12.1 ; sys_platform == 'darwin'",
    "keyring>=25.7.0",
]

[dependency-groups]
dev = [
    "pyinstaller==6.18.0",
    "pytest>=8.0,<9.0",
    "pytest-asyncio>=0.23,<1.0",
    "pyright>=1.1.407",
    "ty>=0.0.9",
    "ruff>=0.14.10",
]

[build-system]
requires = ["uv_build>=0.8.5,<0.10.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "biofoundry_cli"
source-exclude = ["src/biofoundry_cli/deps/**/*"]

[tool.uv.workspace]
members = [
    "packages/biofoundry_llm",
    "packages/biofoundry_os",
]

[tool.uv.sources]
biofoundry_llm = { workspace = true }
biofoundry_os = { workspace = true }
biofoundry-cli = { workspace = true }

[project.scripts]
biofoundry = "biofoundry_cli.cli:cli"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle
    "F",   # Pyflakes
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "I",   # isort
]

[tool.ruff.lint.per-file-ignores]

[tool.pyright]
typeCheckingMode = "standard"
pythonVersion = "3.13"
include = [
    "src/**/*.py",
]
strict = ["src/biofoundry_cli/**/*.py"]

[tool.ty.environment]
python-version = "3.13"

[tool.ty.src]
include = [
    "src/**/*.py",
]

[tool.typos.files]
extend-exclude = ["biofoundry.spec", "pyinstaller.py"]

[tool.typos.default.extend-words]
datas = "datas"
Seeked = "Seeked"
seeked = "seeked"
iterm = "iterm"
